Refactor Cloudflare Worker authentication logic
- Extracted authentication logic to `src/auth.ts`
- Created `authenticateRequest` helper to unify secret and session checks
- Updated `index.ts` to use new helper
- Updated `security.test.ts` to import from `src/auth.ts`
- Verified with typecheck and tests
